+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Release Date: TBD
 - TBS driver compilation failing on RPi OS Bullseye due to disabled SAA7146.
 - Flat panel and Sat-IP antenna/receiver options on the configuration menu
   after C-band satellite selection.
+- Standalone demodulator configuration not stopping on SNMP error.
 
 ## 0.4.4